requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
gunicorn==20.1.0
numpy==1.22.0
pandas==1.3.5
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
